backend/requirements.txt
python-dotenv
google-generativeai
tabulate
flask
flask-cors